Refeactoring object queries
