+++ src/GraphQL/loaders/objectsQueries.js
@@ -317,31 +317,4 @@ const load = parseGraphQLSchema => {
   };
 };
 
-const extractKeysAndInclude = selectedFields => {
-  selectedFields = selectedFields.filter(
-    field => !field.includes('__typename')
-  );
-  let keys = undefined;
-  let include = undefined;
-  if (selectedFields && selectedFields.length > 0) {
-    keys = selectedFields.join(',');
-    include = selectedFields
-      .reduce((fields, field) => {
-        fields = fields.slice();
-        let pointIndex = field.lastIndexOf('.');
-        while (pointIndex > 0) {
-          const lastField = field.slice(pointIndex + 1);
-          field = field.slice(0, pointIndex);
-          if (!fields.includes(field) && lastField !== 'objectId') {
-            fields.push(field);
-          }
-          pointIndex = field.lastIndexOf('.');
-        }
-        return fields;
-      }, [])
-      .join(',');
-  }
-  return { keys, include };
-};
-
-export { getObject, findObjects, load, extractKeysAndInclude };
+export { getObject, findObjects, load };
